#108 Separate out examples
