+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -
 
-## [0.3.0] - 2023-09-??
+## [0.3.0] - 2023-10-01
 
 ### Changed in 0.3.0
