+++ Makefile.darwin
@@ -31,12 +31,12 @@ build: darwin/amd64
 
 .PHONY: test
 test:
-#	@go test -exec macos_exec_dyld.sh -v -p 1 ./...
-	@go test -exec macos_exec_dyld.sh -v ./.
-	@go test -exec macos_exec_dyld.sh -v ./g2config
-	@go test -exec macos_exec_dyld.sh -v ./g2configmgr
-	@go test -exec macos_exec_dyld.sh -v ./g2diagnostic
-	@go test -exec macos_exec_dyld.sh -v ./g2engine
+	@go test -exec macos_exec_dyld.sh -v -p 1 ./...
+#	@go test -exec macos_exec_dyld.sh -v ./.
+#	@go test -exec macos_exec_dyld.sh -v ./g2config
+#	@go test -exec macos_exec_dyld.sh -v ./g2configmgr
+#	@go test -exec macos_exec_dyld.sh -v ./g2diagnostic
+#	@go test -exec macos_exec_dyld.sh -v ./g2engine
 #	@go test -exec macos_exec_dyld.sh -v ./g2product
 
 # -----------------------------------------------------------------------------
